% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pulsar.R
\name{print.Pulsar}
\alias{print.Pulsar}
\title{Print details about a Pulsar instance}
\usage{
\method{print}{Pulsar}(obj)
}
\arguments{
\item{obj}{a Pulsar connection instance}
}
\description{
Print details about a Pulsar instance
}
