% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pulsar.R
\docType{data}
\name{Pulsar}
\alias{Pulsar}
\title{Pulsar class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Pulsar
}
\arguments{
\item{connection}{Pulsar host and port (format: "hostname:port")}

\item{topic}{Full name of the topic including persistence, tenant, namespace and pub/sub name}

\item{onMessageFct}{Function invoked when when receiving message}
}
\value{
a Pulsar connection object
}
\description{
Pulsar class
}
\examples{
\dontrun{
producer <- Pulsar$new("localhost:8080", "persistent/public/default/my-topic")
producer$send("hello")

onMsg = function(payload) {
  cat("Received string:", rawToChar(payload), "\\n")
  TRUE
}
consumer <- Pulsar$new("localhost:8080", "persistent/public/default/my-topic/my-sub",
                       onMessageFct = onMsg)
}
}
\keyword{datasets}
